Provide a framework for creating a NuoDB client package
This framework allows new client packages to be installed by adding a
new client/pkg/*.py file describing how that package should be
downloaded and what elements of the package should be staged, where.

Multiple clients can be extracted from each package, if needed.

A command-line utility "build" controls the creation of the client
package.

A README/manifest file is automatically generated based on the content
of each client.

A client-centric version of nuodb_setup.sh is provided which uses the
XDG standard locations for storing configuration, logging, etc.