pynuoadmin: Pull in the argcomplete module and nuocmd-complete
The 'nuocmd' utility supports command-line completion in bash if you
have installed the Python argcomplete module and configured it in your
shell.

Rather than requiring users to download argcomplete themselves, include
it in the client package.  Also include the nuocmd-complete script from
the database package that can be used to configure the shell properly.