Make publishing easier and introduce bundles
= Generate README.md

Generate package/README.md which can be copy/pasted into the release
notes on GitHub when generating a new client package release.

This ensures that we list the correct version of each software
artifact included in the package, and lists the packages in a
categorized way to improve readability (see below).

= Introduce the concept of a "bundle"

The client package is composed of software extracted from _packages_.

Each package consists of one or more _stages_.

Each _stage_ describes a software artifact that can be extracted from
its downloaded package.

Each _stage_ is associated with exactly one _bundle_, which categorizes
the type of software provided by the _stage_ (e.g., a CLI tool).

Bundles are used in the following ways:

* Categorize each software artifact in the new `README.md`.

* If the option --separate-bundles is supplied, one tarball per bundle
  will be generated (instead of the default which stores all stages
  together into one tarball).

See comments in build and bundles defined in client/bundles.py.

= Add argument --version

Instead of auto-generating a version from the current date, the user
must now supply the version to use with the --version argument.

This avoids the problem where the next version number is communicated
but then the actual package is built on a different day.

= Add argument --separate-bundles

By default, one tarball called nuodb-tools-{version}.{arch}.tar.gz is
created. This package contains the software artifacts from all stages.

If --separate-bundles is supplied, then one tarball is created
_per bundle_. Each tarball contains only the software artifacts from
the stages associated with that bundle.

For example, given the stages and bundles defined as of this commit,
building for Linux without --separate-bundles produces this:

  $ ./build --clean ; ./build --version 2023.1
  $ ls -1 package/*.tar.gz
  package/nuodb-tools-2023.1.lin-x64.tar.gz

whereas building _with_ --separate-bundles produces this:

  $ ./build --clean ; ./build --version 2023.1 --separate-bundles
  $ ls -1 package/*.tar.gz
  package/nuodb-cli-tools-2023.1.lin-x64.tar.gz
  package/nuodb-drivers-2023.1.lin-x64.tar.gz

= Other

* Some trivial renaming of stages.

* Updated the GitHub README.rst with instructions for building a
  client package yourself.

* Added support for a "friendly URL" for each package so that
  a link to a stage's GitHub / Maven Central / PyPi / etc. page
  can be provided in the generated README.md.