Change migrator package to use TAR
With the latest release of migrator tools, we made
available both ZIP and TAR package.

The way the Github metadata class handles packages
is that it takes the first one available.

On previous versions where there was only a ZIP package,
the fetch would work fine, but now the TAR package
has become the first one in the metadata list.

This commit changes the migrator packaging in order to
use TAR package instead of a ZIP package.